segment logic for clickhouse queries that have a lot
