+++ ee/clickhouse/queries/__init__.py
@@ -1,2 +1,2 @@
 from .clickhouse_funnel import ClickhouseFunnel
-from .clickhouse_trends import ClickhouseTrends
+from .trends.clickhouse_trends import ClickhouseTrends
